requirements.txt
# requirements.txt 
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Document Processing
docling==2.8.1
pypdf2==3.0.1

# RAG & Vector Store
langchain==0.0.354
langchain-community==0.0.10
langchain-google-genai==0.0.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Embeddings
torch==2.1.0
transformers==4.35.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0